core_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile core_requirements.in
#
aniso8601==8.0.0          # via flask-restful
click==7.1.2              # via flask
cobble==0.1.3             # via mammoth
flask-restful==0.3.8      # via -r core_requirements.in
flask==1.1.2              # via -r core_requirements.in, flask-restful
gunicorn==20.0.4          # via -r core_requirements.in
html2text==2020.1.16      # via -r core_requirements.in
importlib-metadata==2.0.0  # via markdown
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
lxml==4.6.0               # via python-docx
mammoth==1.4.12           # via -r core_requirements.in
markdown==3.3.1           # via -r core_requirements.in
markupsafe==1.1.1         # via jinja2
pypdf2==1.26.0            # via textspitter
python-decouple==3.3      # via -r core_requirements.in
python-docx==0.8.10       # via textspitter
python-http-client==3.3.1  # via sendgrid
pytz==2020.1              # via flask-restful
sendgrid==6.4.7           # via -r core_requirements.in
simplegeneric==0.8.1      # via -r core_requirements.in
six==1.15.0               # via flask-restful
starkbank-ecdsa==1.1.0    # via sendgrid
textspitter==0.3.2        # via -r core_requirements.in
werkzeug==1.0.1           # via flask
whitenoise==5.2.0         # via -r core_requirements.in
xmljson==0.2.1            # via -r core_requirements.in
zipp==3.3.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
